Gathered all month in same view for calendar
This commit also contains refactorings and some new colors.

+++ SchemeApp/SchemeApp/MonthCollectionViewController.h
@@ -11,7 +11,7 @@
 
 @interface MonthCollectionViewController : UICollectionViewController
 
-@property int year;
-@property int month;
+@property NSInteger year;
+@property NSInteger month;
 
 @end
